Shapes now will be drawn using the lower left as origin
Keep in mind that the drawing origin now is the lower left BUT the shape origin might be the upper left.
If you take a Rectangle the x,y coordinates are still upper left of the Rectangle but will be drawn in the coordinate system with the origin lower left!